+++ R/SCMvextract.R
@@ -1,22 +1,25 @@
-#' @title ASEGvextract
+#' @title SCMvextract
 #'
-#' @description Extracts surface-based measures from ASeg subcortical segmentations in each subject. These must be specifically computed from a FreeSurfer output subjects directory using the python toolbox "SubCortexMesh". ASEGvextract() reads through the metrics .vtk files outputted by SubCortexMesh (surface_metrics/ directory), extracts the vertex-wise values, and stores them as a .rds file, merging left and right hemispheres when applicable, for each (selected) subcortical region. 
-#' @details The function looks for standard subject folder names (starting "sub-") and surface file names as defined by SubCortexMesh (starting "left-","right-", or "brain-stem"). ASEGvextract() requires reticulate and python to read the .vtk files. 
+#' @description Extracts surface-based measures from the python toolbox "SubCortexMesh" in each subject. These can be based on FreeSurfer ASeg or FSL FIRST subcortical segmentations. SCMvextract() reads through the metrics .vtk files outputted by SubCortexMesh (surface_metrics/ directory), extracts the vertex-wise values, and stores them as a .rds file, merging left and right hemispheres when applicable, for each (selected) subcortical region. 
+#' @details The function looks for standard subject folder names (starting "sub-") and surface file names as defined by SubCortexMesh (starting "left-","right-", or "brain-stem"). SCMvextract() requires reticulate and python to read the .vtk files. 
 #' 
 #' Number of vertices in the (bilateral) matrix for each region-of-interest:
-#'  - 2044: accumbens area
-#'  - 3430: amygdala
-#'  - 6940: caudate nuclei
-#'  - 39214: cerebellum
-#'  - 8132: hippocampus
-#'  - 3200: pallidum
-#'  - 8394: putamen
-#'  - 7768: thalamus
-#'  - 7144: ventral diencephalon
-#'  - 9452: brain stem
-#'  - 95718: all aseg ROIs
+#'  - fsaverage/fslfirst: ROI
+#'  - 2044/2026: accumbens area
+#'  - 3430/3592: amygdala
+#'  - 6940/7570: caudate nuclei
+#'  - 39214/31466: cerebellum
+#'  - 8132/8244: hippocampus
+#'  - 3200/3548: pallidum
+#'  - 8394/7908: putamen
+#'  - 7768/8542: thalamus
+#'  - 7144/NA: ventral diencephalon
+#'  - 9452/9516: brain stem
+#'  - 95718/82412: all ROIs merged
+#'  
 #' @param sdirpath A string object containing the path to the 'SubCortexMesh' surface metrics directory, containing surface-based metrics (.vtk files). Default is the current working directory ("./").
 #' @param outputdir A string object containing the path of the directory where all ROI RDS files will be stored. Default is 'subcortices' in the R temporary directory (tempdir()).
+#' @param template A string object containing the name of the template segmentation which was used in SubCortexMesh ('fsaverage' or 'fslfirst').
 #' @param measure A string object containing the name of the measure of interest. Options include thickness, and surfarea. Default is thickness.
 #' @param roilabel A string object or vector of string objects containing the name(s) of the subcortical region(s) to extract (both hemispheres automatically included when applicable). E.g. "thalamus" or c("thalamus", "caudate"). Default is all regions in sdirpath.
 #' @param subj_ID A logical object to determine whether to return a list object containing both subject ID and data matrix.
@@ -25,13 +28,12 @@
 #'
 #' @returns A directory containing - for each bilateral subcortical region separately - a .RDS files, each with a list containing 1. the list of subject IDs (first element) and 2. a surface data matrix object (second element), or only a data matrix object. Each matrix has N subjects x M vertices dimensions and can be used readily by VertexWiseR statistical analysis functions. Each row corresponds to a subject (in the order they are listed in the folder) and contains the left to right hemispheres' vertex-wise values (if applicable).
 #' @examples
-#' ASEGvextract(sdirpath = "subcortexmesh_output_metrics", 
-#' outputdir=paste0(tempdir(), "\\subcortices"), measure="surfarea") 
+#' SCMvextract(sdirpath = "subcortexmesh_output_metrics", 
+#' outputdir=paste0(tempdir(), "\\subcortices"), template='fsaverage', measure="surfarea") 
 #' @importFrom reticulate import
 #' @importFrom stringr str_extract
-#' @export
-
-ASEGvextract=function(sdirpath="./", outputdir, measure = 'thickness', roilabel, subj_ID = TRUE, silent=FALSE, VWR_check=TRUE) 
+#' @export 
+SCMvextract=function(sdirpath="./", outputdir, template, measure = 'thickness', roilabel, subj_ID = TRUE, silent=FALSE, VWR_check=TRUE) 
 { 
 
   #Check required python dependencies. If files missing:
@@ -75,13 +77,17 @@ ASEGvextract=function(sdirpath="./", outputdir, measure = 'thickness', roilabel,
   
   #Brainstem, if computed, added separately as no hemispheric split
   bstem.filelist=list.files(path = sdirpath, pattern=paste0("brain-stem.+",measure),recursive=T)
-  #All aseg, similarly already completely merged
-  allaseg.filelist=list.files(path = sdirpath, pattern=paste0("allaseg.+",measure),recursive=T)
+  #All ROIs, similarly already completely merged
+  if(template=='fsaverage'){allmerged='allaseg'}
+  if(template=='fslfirst'){allmerged='allfslfirst'}
+  allmerged.filelist=list.files(path = sdirpath, 
+                             pattern = paste0(allmerged,".+", measure),
+                              recursive=T)
   
   #subcortical ROIs list
   subcortical_list=unique(stringr::str_extract(lh.filelist, paste0("(?<=left-).+?(?=_", measure, ")")))
   if(length(bstem.filelist)>0) {subcortical_list=append(subcortical_list, 'brain-stem')}
-  if(length(allaseg.filelist)>0) {subcortical_list=append(subcortical_list, 'allaseg')}
+  if(length(allmerged.filelist)>0) {subcortical_list=append(subcortical_list, allmerged)}
   
   #reduced to the roilabel if specified
   if (!missing(roilabel))
@@ -117,21 +123,39 @@ ASEGvextract=function(sdirpath="./", outputdir, measure = 'thickness', roilabel,
   }
   
   #prepare matrices for each subcortices separately
-  #precalculated from fsaverage's aseg converted to surfaces in subcortexmesh
-  aseg_matrices=list(
-    "accumbens-area"=matrix(NA, nrow=NROW(sublist), ncol=2044),
-    "amygdala"=matrix(NA, nrow=NROW(sublist), ncol=3430),
-    "caudate"=matrix(NA, nrow=NROW(sublist), ncol=6940),
-    "cerebellum-cortex"=matrix(NA, nrow=NROW(sublist), ncol=39214),
-    "hippocampus"=matrix(NA, nrow=NROW(sublist), ncol=8132),
-    "pallidum"=matrix(NA, nrow=NROW(sublist), ncol=3200),
-    "putamen"=matrix(NA, nrow=NROW(sublist), ncol=8394),
-    "thalamus"=matrix(NA, nrow=NROW(sublist), ncol=7768),
-    "ventraldc"=matrix(NA, nrow=NROW(sublist), ncol=7144),
-    "brain-stem"=matrix(NA, nrow=NROW(sublist), ncol=9452),
-    "allaseg"=matrix(NA, nrow=NROW(sublist), ncol=95718)
+  #precalculated from SCM's template data surfaces
+  if (missing(template)) {
+    stop('A template argument must be provided ("fsaverage" or "fslfirst").')
+  }
+  if (template=='fsaverage'){
+    scm_matrices=list(
+      "accumbens-area"=matrix(NA, nrow=NROW(sublist), ncol=2044),
+      "amygdala"=matrix(NA, nrow=NROW(sublist), ncol=3430),
+      "caudate"=matrix(NA, nrow=NROW(sublist), ncol=6940),
+      "cerebellum-cortex"=matrix(NA, nrow=NROW(sublist), ncol=39214),
+      "hippocampus"=matrix(NA, nrow=NROW(sublist), ncol=8132),
+      "pallidum"=matrix(NA, nrow=NROW(sublist), ncol=3200),
+      "putamen"=matrix(NA, nrow=NROW(sublist), ncol=8394),
+      "thalamus"=matrix(NA, nrow=NROW(sublist), ncol=7768),
+      "ventraldc"=matrix(NA, nrow=NROW(sublist), ncol=7144),
+      "brain-stem"=matrix(NA, nrow=NROW(sublist), ncol=9452),
+      "allaseg"=matrix(NA, nrow=NROW(sublist), ncol=95718)
     )
-  
+  } else if (template=='fslfirst') {
+    scm_matrices=list(
+      "accumbens-area"=matrix(NA, nrow=NROW(sublist), ncol=2026),
+      "amygdala"=matrix(NA, nrow=NROW(sublist), ncol=3592),
+      "caudate"=matrix(NA, nrow=NROW(sublist), ncol=7570),
+      "cerebellum-cortex"=matrix(NA, nrow=NROW(sublist), ncol=31466),
+      "hippocampus"=matrix(NA, nrow=NROW(sublist), ncol=8244),
+      "pallidum"=matrix(NA, nrow=NROW(sublist), ncol=3548),
+      "putamen"=matrix(NA, nrow=NROW(sublist), ncol=7908),
+      "thalamus"=matrix(NA, nrow=NROW(sublist), ncol=8542),
+      "brain-stem"=matrix(NA, nrow=NROW(sublist), ncol=9516),
+      "allfslfirst"=matrix(NA, nrow=NROW(sublist), ncol=82412)
+    )
+  } else {stop(paste0('Template "', template, '" unknown. Applicable templates are fsaverage or fslfirst. See SubCortexMesh\'s documentation.'))}
+
   #extract vtk scalars for each subject and subcortical region
   for (sub in 1:NROW(sublist))
   {
@@ -140,32 +164,32 @@ ASEGvextract=function(sdirpath="./", outputdir, measure = 'thickness', roilabel,
     lh.filelist.sub=lh.filelist[grep(sublist[sub],lh.filelist)]
     rh.filelist.sub=rh.filelist[grep(sublist[sub],rh.filelist)]
     bstem.filelist.sub=bstem.filelist[grep(sublist[sub],bstem.filelist)]
-    allaseg.filelist.sub=allaseg.filelist[grep(sublist[sub],allaseg.filelist)]
+    allmerged.filelist.sub=allmerged.filelist[grep(sublist[sub],allmerged.filelist)]
       
     for (vol in 1:NROW(subcortical_list))
     {
       vol_label=subcortical_list[vol]
       if (silent==FALSE){message(paste0('=> ', vol_label))}
       
-      if (vol_label!='brain-stem' & vol_label!='allaseg')
+      if (! vol_label %in% c('brain-stem',allmerged))
       {
         if (length(grep(vol_label,lh.filelist.sub)) != 0 ){
         lh_vtk_path=paste0(sdirpath, lh.filelist.sub[grep(vol_label,lh.filelist.sub)])
         lh=vtk_getscalar(lh_vtk_path)
         rh_vtk_path=paste0(sdirpath, rh.filelist.sub[grep(vol_label,rh.filelist.sub)])
         rh=vtk_getscalar(rh_vtk_path)
-        aseg_matrices[[vol_label]][sub,]=c(lh,rh)
+        scm_matrices[[vol_label]][sub,]=c(lh,rh)
         }
       } else if (vol_label=='brain-stem')
       {
         if (length(bstem.filelist.sub) != 0 ){
         bstem_path=paste0(sdirpath,bstem.filelist.sub[grep(vol_label,bstem.filelist.sub)])
-        aseg_matrices[[vol_label]][sub,]=vtk_getscalar(bstem_path)}
-      } else if (vol_label=='allaseg')
+        scm_matrices[[vol_label]][sub,]=vtk_getscalar(bstem_path)}
+      } else if (vol_label == allmerged)
       {  
-        if (length(allaseg.filelist.sub) != 0 ){
-        allaseg_path=paste0(sdirpath,allaseg.filelist.sub[grep(vol_label,allaseg.filelist.sub)])
-        aseg_matrices[[vol_label]][sub,]=vtk_getscalar(allaseg_path)}
+        if (length(allmerged.filelist.sub) != 0 ){
+        allmerged_path=paste0(sdirpath,allmerged.filelist.sub[grep(vol_label,allmerged.filelist.sub)])
+        scm_matrices[[vol_label]][sub,]=vtk_getscalar(allmerged_path)}
       }
     }   
   }
@@ -173,33 +197,33 @@ ASEGvextract=function(sdirpath="./", outputdir, measure = 'thickness', roilabel,
   if (silent==FALSE){message(paste0("Saving data to ", outputdir,"..."))}
   
   #output each ROI in its separate RDS matrix
-  for (surf_dat in 1:length(aseg_matrices))
+  for (surf_dat in 1:length(scm_matrices))
   {
-    if (!all(is.na(aseg_matrices[[surf_dat]])))
+    if (!all(is.na(scm_matrices[[surf_dat]])))
     {
-      surf_obj=aseg_matrices[[surf_dat]]
+      surf_obj=scm_matrices[[surf_dat]]
       
       if (subj_ID == TRUE) 
       {
         row.names(surf_obj)=sublist
-        aseg_matrix=list(surf_obj=surf_obj, sub_list=sublist)
+        scm_matrix=list(surf_obj=surf_obj, sub_list=sublist)
         #remove subjects with no data
-        allna=which(apply(aseg_matrix[[1]], 1, 
+        allna=which(apply(scm_matrix[[1]], 1, 
                           function(x) all(is.na(x))))
         if (length(allna) > 0){
-          aseg_matrix[[1]]=aseg_matrix[[1]][-allna,]
-          aseg_matrix[[2]]=aseg_matrix[[2]][-allna]
+          scm_matrix[[1]]=scm_matrix[[1]][-allna,]
+          scm_matrix[[2]]=scm_matrix[[2]][-allna]
         }
         #update object in the grand list to be returned 
-        aseg_matrices[[surf_dat]]=aseg_matrix
+        scm_matrices[[surf_dat]]=scm_matrix
       } else
       {
         allna=which(apply(surf_obj, 1, function(x) all(is.na(x))))
-        if (length(allna) > 0) {aseg_matrix=surf_obj[-allna,]}
+        if (length(allna) > 0) {scm_matrix=surf_obj[-allna,]}
       }
       
       #save individual object separately
-      saveRDS(aseg_matrix, file=paste0(outputdir,"/",tolower(names(aseg_matrices[surf_dat])),"_", measure,".rds"))
+      saveRDS(scm_matrix, file=paste0(outputdir,"/",tolower(names(scm_matrices[surf_dat])),"_", measure,".rds"))
     }
   }
   
@@ -207,9 +231,10 @@ ASEGvextract=function(sdirpath="./", outputdir, measure = 'thickness', roilabel,
   
   #drop empty matrices
   missing_data=c()
-  for (i in 1:length(aseg_matrices))
-  {if (all(is.na(aseg_matrices[[i]]))) {missing_data=c(missing_data,i)}}
-  if(!is.null(missing_data)){aseg_matrices=aseg_matrices[-missing_data]}
+  for (i in 1:length(scm_matrices))
+  {if (all(is.na(scm_matrices[[i]]))) {missing_data=c(missing_data,i)}}
+  if(!is.null(missing_data)){scm_matrices=scm_matrices[-missing_data]}
   
-  return(aseg_matrices)
+  return(scm_matrices)
 }
+