+++ NEWS.md
@@ -1,4 +1,13 @@
-# VertexWiseR v1.5.0 (TBC)
+# VertexWiseR v1.5.1
+
+## NEW FEATURES
+* VertexWiseR can now analyse subcortical surfaces from SubCortexMesh in both the fsaverage ASeg template and fslfirst template as provided by the python toolbox. Template data required can be downloaded separately according to which template is used. ASEGvextract() is now SCMvextract() as it covers more than FreeSurfer's segmentations.
+
+## FIXES
+* There was a mismatch in VertexWiseR 1.5.0's template data and SubCortexMesh's template data for the allaseg surface, which has been reordered since to standardize across FreeSurfer and FSL FIRST-based SCM outputs. This is now fixed.
+* plot_overlay_surf's custom limits did not work properly with the NULL option breaking the plotter. This is now fixed
+
+# VertexWiseR v1.5.0
 
 ## NEW FEATURES
 * VertexWiseR can now analyse surface-based metrics for FreeSurfer's ASeg subcortical regions, converted to surface meshes via the python module [SubCortexMesh](https://github.com/chabld/SubCortexMesh), and extracted with the new R function ASEGvextract(). A new example tutorial was created as a demonstration:  "Example analyses with VertexWiseR - Example 3".
@@ -8,7 +17,7 @@
 ## FIXES
 * In VWRfirstrun(), the check and message prompt for the neurosynth database now prints the path in which it is meant to be stored, instead of an empty "()".
 * The model_check() internal function inside the modelling functions could fail if 'model' or 'contrast' were given as a one-column data.frame instead of a simple vector, as the inherits() checks would not detect numeric or character classes.
-* plot_surf() (with show.plot.window=TRUE) will no longer plot on top of the previous plot in the RStudio "Plots" tab when running the function multiple times, and reset it first.
+* plot_surf() (with show.plot.window=TRUE) will no longer plot on top of the previous plot in the RStudio "Plots" tab when running the function multiple times, and will reset it first.
 * plot_overlay_surf() now accepts a custom limits_2 instead of failing. 
 * Clearer error message for plot_surf() when provided with surf_data that only contains NA values, and if surf_data has multiple rows, empty rows are simply not plotted.
 * Clearer error management and messaging when the title argument is wrongly inputted in plot_surf() (e.g. if there are more titles than there are surface rows).